// ui/types/generated/Node.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Node = {
  id: string;
  vaultId: string;
  subVaultId: string | null;
  nodeType: string;
  title: string;
  summary: string;
  detail: string | null;
  source: string | null;
  sourceType: string | null;
  privacyTier: string | null;
  priority: string;
  version: number;
  isArchived: boolean;
  createdAt: string;
  updatedAt: string;
  lastAccessed: string;
  deletedAt: string | null;
  meta: string;
};